Add missing (GET) params to OAuth2Client.request
When calling the `request` method with `params` dict, the dict is not converted
to GET parameters (e.g. `params={"foo": "bar"}` to `f"{url}?foo=bar"`).
Current workaround is to include it in the url.

This is caused by the method `request` in class `OAuth2Client` not passing on
the option `params`. One way to fix this would be to pass it implicitly by
removing `params: t.Dict = None` from the options.

I opted for the more explicit fix and passing `params=params` to the `_request`
method.